refactor: move build_tasks_table_component into layouts/case_detail.py
